+++ examples/sync.rs
@@ -1,7 +1,0 @@
-use online::sync::check;
-
-fn main() {
-    println!("Online? {}", check(None).is_ok());
-
-    println!("Online (`Result`)? {:?}", check(None).unwrap());
-}